refactor(logging): Se actualizan las importaciones de logging para utilizar la instancia centralizada CustomLogger en shared
