+++ backend/java-server/src/main/java/com/giozar04/bankClients/domain/models/BankClientRepositoryAbstract.java
@@ -6,16 +6,15 @@ import java.util.Objects;
 import com.giozar04.bankClient.domain.entities.BankClient;
 import com.giozar04.bankClients.domain.interfaces.BankClientRepositoryInterface;
 import com.giozar04.databases.domain.interfaces.DatabaseConnectionInterface;
-import com.giozar04.shared.logging.CustomLogger;
+import com.giozar04.logging.CustomLogger;
 
 public abstract class BankClientRepositoryAbstract implements BankClientRepositoryInterface {
 
     protected final DatabaseConnectionInterface databaseConnection;
-    protected final CustomLogger logger;
+    protected final CustomLogger logger = CustomLogger.getInstance();
 
     protected BankClientRepositoryAbstract(DatabaseConnectionInterface databaseConnection) {
         this.databaseConnection = Objects.requireNonNull(databaseConnection, "La conexión no puede ser nula");
-        this.logger = new CustomLogger();
     }
 
     protected void validateBankClient(BankClient client) {
