+++ backend/java-server/src/main/java/com/giozar04/users/domain/models/UserRepositoryAbstract.java
@@ -4,19 +4,18 @@ import java.util.List;
 import java.util.Objects;
 
 import com.giozar04.databases.domain.interfaces.DatabaseConnectionInterface;
-import com.giozar04.shared.logging.CustomLogger;
+import com.giozar04.logging.CustomLogger;
 import com.giozar04.users.domain.entities.User;
 import com.giozar04.users.domain.interfaces.UserRepositoryInterface;
 
 public abstract class UserRepositoryAbstract implements UserRepositoryInterface {
 
     protected final DatabaseConnectionInterface databaseConnection;
-    protected final CustomLogger logger;
+    protected final CustomLogger logger =  CustomLogger.getInstance();
 
     protected UserRepositoryAbstract(DatabaseConnectionInterface databaseConnection) {
         this.databaseConnection = Objects.requireNonNull(databaseConnection, 
             "La conexión a la base de datos no puede ser nula");
-        this.logger = new CustomLogger();
     }
 
     protected void validateUser(User user) {
